api/requirements.txt
fastapi==0.110.0
uvicorn==0.30.1
langchain-core==0.2.36
langchain-community==0.2.16
huggingface-hub==0.22.2
supabase==2.5.0
python-dotenv==1.0.1
pydantic==2.9.2
